requirements.txt
bcrypt==3.1.7
blinker==1.4
cffi==1.14.0
Click==7.0
client==0.0.1
coverage==4.4
cryptography==2.8
Flask==1.1.1
Flask-Bcrypt==0.7.1
flask-buzz==0.1.14
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
Flask-Mail==0.9.1
flask-marshmallow==0.10.1
flask-praetorian==1.0.0
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
marshmallow==3.3.0
marshmallow-sqlalchemy==0.21.0
ndg-httpsclient==0.5.1
okpy==1.14.19
passlib==1.7.2
pendulum==2.0.5
psycopg2==2.8.4
pusher==2.1.4
py-buzz==0.3.7
pyasn1==0.4.8
pycparser==2.19
PyJWT==1.7.1
PyNaCl==1.3.0
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-dotenv==0.10.5
pytzdata==2019.3
requests==2.12.4
six==1.14.0
SQLAlchemy==1.3.13
urllib3==1.25.8
Werkzeug==0.16.0
